don't die on database lock
Sometimes it's useful to query the database of a running application.
Until now, that would potentially cause the server to crash.
Now the server will wait until the lock is no longer held.

To test this, you can lock the DB like so:

    SELECT listings.id FROM listings CROSS JOIN listings as L1;

With enough listings, this takes a few seconds.